Rename all packages to be under org/antlr/v5 and not com/strumenta
