+++ runtime-testsuite/test/org/antlr/v5/test/runtime/kotlin/commonTest/org/antlr/v4/kotlinruntime/StringCharStreamTest.kt
@@ -5,7 +5,7 @@
  */
 package org.antlr.v5.test.runtime.kotlin.commonTest.org.antlr.v5.kotlinruntime
 
-import com.strumenta.antlrkotlin.runtime.ext.codePointIndices
+import org.antlr.v5.kotlinruntime.ext.codePointIndices
 import org.antlr.v5.kotlinruntime.CharStreams
 import org.antlr.v5.kotlinruntime.misc.Interval
 import org.junit.jupiter.api.Test
