+++ runtime-testsuite/test/org/antlr/v5/test/runtime/kotlin/helpers/RuntimeTestPrintStream.kt
@@ -1,5 +1,7 @@
 package org.antlr.v5.test.runtime.kotlin.helpers
 
+import org.antlr.v5.kotlinruntime.System
+
 class RuntimeTestPrintStream(outStream: java.io.OutputStream) :
     java.io.PrintStream(outStream),
-    com.strumenta.antlrkotlin.runtime.System.PrintStream
+    System.PrintStream
