+++ runtime/Kotlin/src/jvm/org/antlr/v5/kotlinruntime/BitSet.kt
@@ -4,7 +4,7 @@
  * can be found in the LICENSE.txt file in the project root.
  */
 
-package com.strumenta.antlrkotlin.runtime
+package org.antlr.v5.kotlinruntime
 
 import java.util.BitSet as JavaBitSet
