+++ runtime/Kotlin/src/jvm/org/antlr/v5/kotlinruntime/Environment.kt
@@ -3,7 +3,10 @@
  * Use of this file is governed by the BSD 3-clause license that
  * can be found in the LICENSE.txt file in the project root.
  */
-package com.strumenta.antlrkotlin.runtime
+package org.antlr.v5.kotlinruntime
 
-public inline fun <R> synchronized(lock: Any, block: () -> R): R =
-  kotlin.synchronized(lock, block)
+import java.lang.System as JavaSystem
+
+@Suppress("NOTHING_TO_INLINE")
+internal inline fun platformGetEnv(name: String): String? =
+  JavaSystem.getenv(name)
