+++ runtime/Kotlin/src/jvm/org/antlr/v5/kotlinruntime/IdentityHashMap.kt
@@ -3,6 +3,9 @@
  * Use of this file is governed by the BSD 3-clause license that
  * can be found in the LICENSE.txt file in the project root.
  */
-package com.strumenta.antlrkotlin.runtime
 
-public typealias CopyOnWriteArrayList<E> = java.util.concurrent.CopyOnWriteArrayList<E>
+package org.antlr.v5.kotlinruntime
+
+import java.util.IdentityHashMap as JavaIdentityHashMap
+
+public typealias IdentityHashMap<K, V> = JavaIdentityHashMap<K, V>
