+++ runtime/Kotlin/src/jvm/org/antlr/v5/kotlinruntime/Synchronized.kt
@@ -3,9 +3,7 @@
  * Use of this file is governed by the BSD 3-clause license that
  * can be found in the LICENSE.txt file in the project root.
  */
+package org.antlr.v5.kotlinruntime
 
-package com.strumenta.antlrkotlin.runtime
-
-import java.util.IdentityHashMap as JavaIdentityHashMap
-
-public typealias IdentityHashMap<K, V> = JavaIdentityHashMap<K, V>
+public inline fun <R> synchronized(lock: Any, block: () -> R): R =
+  kotlin.synchronized(lock, block)
