+++ runtime/Kotlin/src/org/antlr/v5/kotlinruntime/Lexer.kt
@@ -5,7 +5,6 @@
  */
 package org.antlr.v5.kotlinruntime
 
-import com.strumenta.antlrkotlin.runtime.System
 import org.antlr.v5.kotlinruntime.atn.LexerATNSimulator
 import org.antlr.v5.kotlinruntime.misc.IntegerStack
 import org.antlr.v5.kotlinruntime.misc.Interval
