+++ runtime/Kotlin/src/org/antlr/v5/kotlinruntime/ProxyErrorListener.kt
@@ -5,7 +5,6 @@
  */
 package org.antlr.v5.kotlinruntime
 
-import com.strumenta.antlrkotlin.runtime.BitSet
 import org.antlr.v5.kotlinruntime.atn.ATNConfigSet
 import org.antlr.v5.kotlinruntime.dfa.DFA
 
@@ -32,13 +31,13 @@ public class ProxyErrorListener(private val delegates: Collection<org.antlr.v5.k
   }
 
   override fun reportAmbiguity(
-    recognizer: Parser,
-    dfa: DFA,
-    startIndex: Int,
-    stopIndex: Int,
-    exact: Boolean,
-    ambigAlts: BitSet,
-    configs: ATNConfigSet,
+      recognizer: Parser,
+      dfa: DFA,
+      startIndex: Int,
+      stopIndex: Int,
+      exact: Boolean,
+      ambigAlts: BitSet,
+      configs: ATNConfigSet,
   ) {
     for (listener in delegates) {
       listener.reportAmbiguity(recognizer, dfa, startIndex, stopIndex, exact, ambigAlts, configs)
@@ -46,12 +45,12 @@ public class ProxyErrorListener(private val delegates: Collection<org.antlr.v5.k
   }
 
   override fun reportAttemptingFullContext(
-    recognizer: Parser,
-    dfa: DFA,
-    startIndex: Int,
-    stopIndex: Int,
-    conflictingAlts: BitSet,
-    configs: ATNConfigSet,
+      recognizer: Parser,
+      dfa: DFA,
+      startIndex: Int,
+      stopIndex: Int,
+      conflictingAlts: BitSet,
+      configs: ATNConfigSet,
   ) {
     for (listener in delegates) {
       listener.reportAttemptingFullContext(recognizer, dfa, startIndex, stopIndex, conflictingAlts, configs)
