+++ runtime/Kotlin/src/org/antlr/v5/kotlinruntime/TokenStreamRewriter.kt
@@ -5,7 +5,6 @@
  */
 package org.antlr.v5.kotlinruntime
 
-import com.strumenta.antlrkotlin.runtime.System
 import org.antlr.v5.kotlinruntime.misc.Interval
 import kotlin.math.max
 import kotlin.math.min
