+++ runtime/Kotlin/src/org/antlr/v5/kotlinruntime/UnbufferedTokenStream.kt
@@ -6,8 +6,6 @@
 
 package org.antlr.v5.kotlinruntime
 
-import com.strumenta.antlrkotlin.runtime.System
-import com.strumenta.antlrkotlin.runtime.assert
 import org.antlr.v5.kotlinruntime.misc.Interval
 import kotlin.math.min
