+++ runtime/Kotlin/src/org/antlr/v5/kotlinruntime/atn/SingletonPredictionContext.kt
@@ -6,7 +6,7 @@
 
 package org.antlr.v5.kotlinruntime.atn
 
-import com.strumenta.antlrkotlin.runtime.assert
+import org.antlr.v5.kotlinruntime.assert
 
 @Suppress("EqualsOrHashCode")
 public open class SingletonPredictionContext internal constructor(
