+++ runtime/Kotlin/src/org/antlr/v5/kotlinruntime/ext/String.ext.kt
@@ -3,7 +3,7 @@
  * Use of this file is governed by the BSD 3-clause license that
  * can be found in the LICENSE.txt file in the project root.
  */
-package com.strumenta.antlrkotlin.runtime.ext
+package org.antlr.v5.kotlinruntime.ext
 
 public fun String.codePointIndices(): IntArray {
   val intArray = IntArray(length)
